+++ WalletAdmin/Mapping/MovimentoSaidaMap.cs
@@ -16,13 +16,13 @@ namespace WalletAdmin.Mapping
             Id(x => x.SAI_CODIGO, x =>
             {
                 x.Generator(Generators.Increment);
-                x.Type(NHibernateUtil.Int64);
+                x.Type(NHibernateUtil.Int32);
                 x.Column("SAI_CODIGO");
             });
 
             Property(b => b.PES_CODIGO, x =>
             {
-                x.Type(NHibernateUtil.Int64);
+                x.Type(NHibernateUtil.Int32);
                 x.NotNullable(true);
             });
             Property(b => b.SAI_DATA, x =>
@@ -37,7 +37,7 @@ namespace WalletAdmin.Mapping
             });
             Property(b => b.SAI_VALOR, x =>
             {
-                x.Type(NHibernateUtil.Double);
+                x.Type(NHibernateUtil.Decimal);
             });
             Table("Tabela_Movimento_Saida");
         }
